fix(install): try to kill zen only if open

+++ install.sh
@@ -71,7 +71,9 @@ install_npm() {
 
 stow_dots() {
   # kill zen to prevent overwriting of files
-  killall zen-bin
+  if pgrep -x "zen-bin" >/dev/null; then
+    killall zen-bin
+  fi
   shopt -s dotglob nullglob
 
   for dir in "$DOTS_DIR"/*; do
